Fix STL warnings

+++ src/stl/algorithm/NonModifying/E012_CountIf/main.cpp
@@ -11,7 +11,7 @@
 int main() {
     std::vector<int> numbers = { 0, 1, 2, 3, 4, 5, 6, 7, 8 , 9 };
 
-    int counter = std::count_if(numbers.begin(), numbers.end(), [](int n) {
+    long counter = std::count_if(numbers.begin(), numbers.end(), [](int n) {
         return n % 2 == 1;
     });
